requirements.txt
beautifulsoup4==4.12.0
Django==4.1.7
djangorestframework==3.14.0
requests==2.28.2
